requirements.txt
loguru
pillow
pyyaml>3.12